In pysandesh library, sandesh message that is dropped should be logged at the level specified by the user in the corresponding sandesh.
pysandesh library logs sandesh messages that are dropped @ SYS_INFO level.
If the sandesh is logged with level SYS_ERR and the python daemon is configured to
write logs in the local file with SYS_ERR, then this sandesh message wouldn't be
written in the log file if this message is dropped. Therefore, sandesh message
that is dropped should be logged at the level specified by the user in the corresponding sandesh.

Change-Id: Ic418e247eadd39962ef0bebc1662c651f5357473
Closes-Bug: #1392932