Option to override the key annotation in sandesh UVE/Alarm at run time
Presently, the value for key annotation need to be specified while
defining the sandesh UVE/alarms in .sandesh
If the same data needs to be used for different table,
then with the current implementation, duplicate structures
(with different name) needs to be defined with different
value for the key annotation.
We plan to use the same schema/structure for all the alarms.
Therefore, it would be useful to have the ability to override
the key annotation at run time so that the same definition of
alarm can be reused.

With this change, table name can be specified during the
instantiation of the sandesh UVE/Alarm. If the table name
is not specified, then the key value specified in the
sandesh definition would be used.

This diff has only pysandesh changes. Need to make corresponding
changes in sandesh c++ library.

Change-Id: I36c2941abc3aabe636722655fe8f193e1389fb58
Partial-Bug: #1420807