Handle event _EV_START in state _IDLE in pysandesh statemachine
Typically, the admin state is enabled/disabled from the test code.
When the admin state is disabled in state _ESTABLISHED, the state
transitions to _IDLE. When the admin state is enabled,
then the event _EV_START is enqueued in the statemachine. However,
the event _EV_START is not handled in state _IDLE.

Change-Id: I8defe05a026070f23359de356a00ccec45ddabb1