Option to override the key annotation in sandesh UVE/Alarm at run time
Presently, the value for key annotation need to be specified while
defining the sandesh UVE/alarms in .sandesh
If the same data needs to be used for different table,
then with the current implementation, duplicate structures
(with different name) needs to be defined with different
value for the key annotation.
This patch adds the capability to override the key annotation at
run time for sandesh uve and alarm in sandesh c++ library. The
corresponding changes for pysandesh library was already committed.

Change-Id: Ia448400169ca8bba6f3c3b53dd7e5a8184f18e3d
Closes-Bug: #1420807