We need to make alarm information more structred and easier for clients to parse. The "description" field in alarms is being deprecated. Instead, we have a "rules" list.
Each rule has an operator (e.g. "!=") , and one or two operands.
Each operand has a json-encoded value (E.g. "PROCESS_RUNNING") ,
and an optional name (e.g. "NodeStatus.process_status")

If we are comparing two UVE attributes, both are operands with names.

If we compare a UVE attribute against a constant, the 1st operand
has the name of the UVE attribute, but the 2nd is a constant and
does not have a name

Change-Id: I287efed6026a79768f106494e6366b2e57d0b5db
Partial-Bug: 1428271