Call 'available_write' only when realloc is needed
A call to any function takes a minute amount of time, which over
a large number of calls (order of 100s of thousands) results in
visible usage time. 'available_write' is called for every write
of encoded sandesh to the buffer that holds the encoding and hence
will leave a significant footprint. Hence, call 'available_write'
only if we know that the the buffer space is less than what is
needed to write.

Also, going through protocol layer from the layer that implements
the specific protocol seems an overkill.

Partial BUG: #1528174
Change-Id: Iaa913f304a562a643d84ad77727513f3d32ec726