Adding non-inline UVE-Cache support, based on the "period" annotation on UVE Structs (This will be used for implementing non-inline derived stats)
For UVE structs that are annotated with "period", the "Send" function only updates the
UVE-Cache (including derived stats); it does not actually send the message.

In this non-inline case, we do NOT load the contents of the derived stats structures
back from the UVE Cache into the UVE at update time (since the UVE is not sent out anyway)

Instead, we only load the contents from the UVE cache to the message when the message
is actually sent out. Right now, the UVE is only sent out during UVE-Sync.
Later, we will also add timers to send the UVE out periodically according to the time
specified in the "period" annotation.

Change-Id: I84173fb0bc73e7c5c1e6e79eb4b082e66e90f87d
Partial-Bug: 1428271