Adding new annotation called 'metric' This will be used to indicate whether a raw attribute is a aggregate counter , diff counter, or a dial. The values are agg,diff and dial DerivedStats will convert aggs into diff while updating DerivedStats, so that further aggregations can be applied to the diff. e.g. EWM The same annotations on both raw attributes and derived stats can also be used in future for generating STATs schemas.
Also, we are adding two new annotations. DSChange can be used by clients to ensure that an attribute is not sent to collector
unless it changes. DSNon0 can be used to ensure that an attribute is not sent to the collector unless it is non-0.

EWM Anomaly detection now does not report its contents if the previous and current values of sigma
are between 0.5 and -0.5. (expect during sync and introspect)

Partial-Bug:1601908

Change-Id: Idd9bbecc0ce2d110f8771ac48e3e01ae1241a9db