python sandesh client library changes to accept more than 2 collectors
Presently, the contrail services may be either provisoned with the
collector list or it gets the collector list from discovery service. In
either case, the sandesh client library doesn't accept more than 2
collectors. With the removal of discovery service, all the contrail services
that connects to collector would be provisioned with the collector list.
Therefore, it is necessary that the sandesh client library accepts
collector list without any limitation.

With this patch, init_generator() can accept collector list of size > 2 and
the notion of primary and secondary collector has been removed. The
sandesh client would connect to the first collector in the list and upon
connection failure/closure would connect to the next collector in the
list and so on.

Added a new function reconfig_collectors() - clients should call this
function upon receiving SIGHUP to indicate change in the collector list
or rebalance the collector connections. It is expected that the sandesh
clients would randomize the collector list before passing it to
init_generator() and reconfig_collectors()

This patch doesn't remove the code that subscribes for the collector
service with discovery as this would break the functionality without the
provisioning/controller changes.

TODO:
- UT will be added in the next commit.

Change-Id: If5fe7f332cb65d1f9fa8489e52f913803ed085d9
Partial-Bug: #1641848