Adding a new DerivedStats interface: DerivedStatsPeriodicAnomalyIf Using this interface, we can aggregate a raw attribute periodically before subjecting the result to Anomaly Detection. E.g. 12: optional derived_stats_results.AnomalyResult sum_ewm_x (stats=1.DSSum-x:DSAnomaly:EWM:0.2) Here, we will be doing a SUM aggregation on the attribute x periodically. This aggregated result will then be fed into the EWM Anomaly detetction algorithm.
Also adding "timeout" capability to periodic UVE processing
If the periodic UVE is not updated for the given number of
processing cycles, it will be automatically deleted.

Adding a new Send interface in UVE cache for supplying proxy and partition.
This allows for bulk deletion of UVEs in the cache on a per proxy-partition basis
UVEAggregation in collector will use this interface, so that when a collector
loses some partitions, it can clear the corresponding Aggregate UVEs.

Closes-Bug:1667013

Change-Id: Ie1f4a7aafa6fb982fb186a37faf705de3cb21eaf