Introspect of daemons missing links to some request files
Global static initialization is used to populate a SandeshHttp
static map with the names of the sandesh IDL files that contain
requests for introspect for a daemon. The static declaration and
definition is done in _html.cpp. In order for the linker to include
the symbols in that translation unit a standalone int is also defined
in the file and used in _types.cpp. However the current usage was just
assigning another static variable this int and as a result the symbols
in the _html.cpp were not included and the map was not populated.
Changed the generated code to assign a value to the standalone int
as part of RequestFromHttp function so that symbols from the
_html.cpp translation units are pulled in.

Change-Id: Id5515bb75fab093fbc62bcebf8fecb5dbed2be9c
Closes-Bug: #1691605
Closes-Bug: #1691949