Teardown sandesh client connection to prevent unbounded queueing
Currently UVEs are not dropped on the sandesh client and
hence in scenarios where the collector is not reading the
messages due to database write backoff it can result in the
sandesh client queue size getting unbounded. Fix is to
teardown the sandesh client connection and force a resync
of the UVEs with the collector to avoid this. Added an
exponential backoff to the teardown logic.

Closes-Bug: #1706487

Conflicts:
	library/python/pysandesh/sandesh_base.py
	library/python/pysandesh/test/sandesh_msg_test.py

Change-Id: I9fff839426d62f8db4fa564c0f1987eeb808fa09