Logging of sessions
This fix lets the session logs go to any or all of collector/
file/syslog. To accommodate this, 2 new logger object is
created for handling only SLO/Sample session destinations. The
destinations to this logger are provided by the agent. A snippet
of the generated_file is added to the bug.
Partial-Bug: 1729812

Change-Id: I8a51aaebd98714c7067c2a942025f7e5ca97dd68