Remove sandesh log that prints change in sending level
This message is logged when the sandesh log level to the collector
changes due to sandesh queue reaching the high/low watermark. This
message is currently logged at level SYS_INFO. This message is not
useful and can flood the log file if the log_level is set to SYS_INFO or
above and hence removed.

Change-Id: Id97368f6101880bc22fd9ee7be03c3674d44d56d
Closes-Bug: #1790737